feat: security i18n
